be/feat: 그룹 1차 완성,TODO 남아있긴함

+++ backend/main-service/src/main/java/com/ssafy/_66days/group/model/dto/GroupSearchPageResponseDTO.java
@@ -19,7 +19,7 @@ public class GroupSearchPageResponseDTO {
     private String name;
     private List<String> categories;
     private String description;
-    private int memberCounts;
+    private Long memberCounts;
     private int maxMemberCounts;
 
     public static GroupSearchPageResponseDTO of(Group group, User user) {
